+++ src/boxyard/cmds/_sync_box.py
@@ -236,6 +236,7 @@ async def sync_box(
                 filters_path=_rclone_filters_path,
                 verbose=verbose,
                 show_rclone_progress=show_rclone_progress,
+                preserve_exec_perms=True,
             )
     
         # Update remote index cache